Update WebAPI project template for beta5
This commit updates WebAPI project template to one coming with beta5:
- new hosting.ini file added to project
- updated project.json
- new README.md based on current html readme document from ASP.NET
- the ValuesController renamed and moved into more proper directory structure
- the tests are updated to new content and updated structure
