+++ templates/projects/webapi/hosting.ini
@@ -1,0 +1,2 @@
+server=Microsoft.AspNet.Server.WebListener
+server.urls=http://localhost:5000
